Insights
Problem Statement :

RSVP Movies is an Indian film production company which has produced many super-hit movies. They have usually released movies for the Indian audience but for their next project, they are planning to release a movie for the global audience in 2022.

The Insights driven from MySQL answers all the problem statements and helps to make the best decision.